requirements.txt
# Core LLM Framework
langchain>=0.1.0
langgraph>=0.0.40
openai>=1.10.0

# Vector Database
qdrant-client>=1.7.0

# UI Framework
streamlit>=1.28.0

# Configuration
python-dotenv>=1.0.0

# Observability
langfuse>=2.0.0

# Utilities
tiktoken>=0.5.0
pandas>=2.0.0

# Document Processing
PyPDF2>=3.0.0
pycryptodome>=3.19.0

# Embeddings & NLP
sentence-transformers>=2.2.0

# Image Processing (for graph visualization)
Pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Optional: For better performance
# watchdog>=3.0.0  # Uncomment for Streamlit auto-reload